fix: ロジックをhookに外だし
