+++ src/pages/_app.tsx
@@ -1,5 +1,5 @@
-import { useEffect } from 'react'
-import { Box, useBreakpointValue, useColorModeValue } from '@chakra-ui/react'
+import { memo } from 'react'
+import { Box, useBreakpointValue } from '@chakra-ui/react'
 import * as Sentry from '@sentry/react'
 import { Integrations } from '@sentry/tracing'
 import { AppProps } from 'next/app'
@@ -7,10 +7,11 @@ import { FooterComponent } from '@/components/organisms/Footer'
 import { HeaderComponent } from '@/components/organisms/Header'
 import { SpHeaderComponent } from '@/components/organisms/SpHeader'
 import Provider from '@/components/templates/Provider'
+import useFillHeightEffect from '@/hooks/useFillHeightEffect'
+import useFontSizeEffect from '@/hooks/useFontSizeEffect'
 import usePageView from '@/hooks/usePageView'
 import { Container } from '@/styles/global.css'
 import '@/styles/globals.css'
-import f from '@/utils/fontSize'
 
 Sentry.init({
   dsn: process.env.NEXT_PUBLIC_SENTRY_DSN,
@@ -22,7 +23,7 @@ Sentry.init({
   tracesSampleRate: 1.0,
 })
 
-const Header = () => {
+const Header = memo(() => {
   const bp = useBreakpointValue({ base: 'base', md: 'md' })
   return (
     <>
@@ -30,49 +31,31 @@ const Header = () => {
       {bp === 'md' && <HeaderComponent />}
     </>
   )
-}
+})
 
 function MyApp({ Component, pageProps }: AppProps) {
-  useEffect(() => {
-    // 文字サイズの取得
-    const fontSize = f.getFontSize()
-    // 文字サイズの保存
-    f.setFontSize(fontSize)
-    // vhの値の保存
-    setFillHeight()
-    // イベントの生成
-    if (window) {
-      window.addEventListener('resize', setFillHeight)
-    }
-    ;() => {
-      // イベントの削除
-      window.removeEventListener('resize', setFillHeight)
-    }
-  }, [])
+  // フォントサイズの大きさの取得 / 設定
+  useFontSizeEffect()
+  // 高さの検知と変更
+  useFillHeightEffect()
   // PVをカウントするイベント
   usePageView()
   return (
     <Provider>
       <Container>
+        {/* ヘッダー */}
         <Header />
+        {/* メインコンテンツ */}
         <Box w="full" as="main" h="100%">
           <Component {...pageProps} />
         </Box>
+        {/* フッター */}
         <Box position={'absolute'} bottom={0} left={0} w="100vw">
           <FooterComponent />
         </Box>
       </Container>
     </Provider>
   )
-}
-
-const setFillHeight = () => {
-  if (window) {
-    // 最初に、ビューポートの高さを取得し、0.01を掛けて1%の値を算出して、vh単位の値を取得
-    let vh = window.innerHeight * 0.01
-    // カスタム変数--vhの値をドキュメントのルートに設定
-    document.documentElement.style.setProperty('--vh', vh + 'px')
-  }
 }
 
 export default MyApp
